requirements.txt
fastapi[all] 
psycopg2-binary
sqlalchemy-utils
alembic
python-dotenv